Added the ability to dig up treasure
Added a "GroundTreasure" prefab that digs up a treasure chest (currently just interactable block placeholder). Added a game manager to keep track of gold count and added helper method to calculate a position in front of the player that is on the grid coordinate system. Refactored shove mechanic to use this helper method